+++ src/app/_component/AuthSession.tsx
@@ -5,9 +5,13 @@ import { ReactNode } from 'react';
 
 type AuthSessionProps = {
   children: ReactNode;
-  initialSession: any; // 서버에서 전달된 초기 세션
+  initialSession?: any; // 서버에서 전달된 초기 세션
 };
 
-export default function AuthSession({ children, initialSession }: AuthSessionProps) {
-  return <SessionProvider session={initialSession}>{children}</SessionProvider>;
+export default function AuthSession({ children }: AuthSessionProps) {
+  return (
+    <SessionProvider >
+      {children}
+    </SessionProvider>
+  );
 }
