+++ src/types/next-auth.d.ts
@@ -5,6 +5,7 @@ declare module 'next-auth' {
   interface Session {
     user?: {
       id?: string;
+      username?: string
     } & DefaultSession['user'];
   }
